+++ src/components/common/Card/Card.css
@@ -1,0 +1,64 @@
+.card-wrapper {
+  position: relative;
+  display: flex;
+  justify-content: center;
+  align-items: flex-start;
+  padding-top: 20px; /* 조명을 카드보다 위에 위치시키기 위한 공간 */
+}
+
+.card {
+  position: relative;
+  display: flex;
+  flex-direction: column;
+  transition: transform 0.3s ease, box-shadow 0.3s ease;
+  z-index: 10; /* 전구보다 아래에 위치하도록 설정 */
+}
+
+.card:hover {
+  transform: scale(1.05);
+}
+
+.light {
+  position: absolute;
+  top: -20px;
+  left: 50%;
+  transform: translateX(-50%);
+  width: 40px;
+  height: 20px;
+  background-color: #ffffff81; /* 전구 색상 */
+  border-radius: 10px 10px 0 0;
+  box-shadow: 0px 0px 15px rgba(255, 250, 204, 0.9),
+    0px 0px 30px rgba(255, 250, 210, 0.7), 0px 0px 45px rgba(248, 240, 162, 0.5);
+  opacity: 0;
+  transition: opacity 0.3s ease;
+  pointer-events: none;
+  z-index: 15; /* 오버레이보다 위에 나타나도록 설정 */
+}
+
+.card-wrapper:hover .light {
+  opacity: 1;
+}
+
+.light::after {
+  content: "";
+  position: absolute;
+  top: 20px;
+  left: 50%;
+  transform: translateX(-50%);
+  width: 350px; /* 조명 확산 크기 */
+  height: 500px;
+  background: radial-gradient(
+    circle,
+    rgba(255, 253, 245, 0.4),
+    transparent 80%
+  );
+  opacity: 0;
+  transition: opacity 0.3s ease;
+  pointer-events: none;
+  filter: blur(12px);
+  z-index: 14; /* 전구보다 약간 아래에 위치 */
+}
+
+.card-wrapper:hover .light::after {
+  opacity: 1;
+}
