+++ src/components/common/SwiperCommentList.jsx
@@ -27,12 +27,11 @@ const SwiperCommentList = ({ title, data }) => {
 
   return (
     <>
-      {/* 평점순 */}
       <Typography
         variant="h4"
         color="white"
         fontWeight={700}
-        sx={{ marginBottom: "1rem" }}
+        sx={{ marginBottom: "20px" }}
       >
         {title}
       </Typography>
